Added RecordList class
RecordList is a List type collection whose elements can be accessed via index operator.
Note: Due to TypeScript limitations, assigning to an index that is out of list's bounds will not throw an error, but the list iterator and several other operations will not work properly.
Avoid using indexed assign as much as possible.